fix debug log code and a few warnings
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@2314 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ platform/windows/orxscrpt/classfactory.cpp
@@ -55,7 +55,7 @@ HRESULT CreateComponentCategory(CATID catid, OLECHAR* catDescription)
   ICatRegister *pcr = NULL;
   HRESULT      hr = S_OK;
   CATEGORYINFO catinfo;
-  int          iLen;
+  size_t       iLen;
 
   hr = CoCreateInstance(CLSID_StdComponentCategoriesMgr,
                         NULL, CLSCTX_INPROC_SERVER, IID_ICatRegister,
